MinipassWebSite/requirements.txt
flask
flask-mail
stripe
python-dotenv
dotenv
bcrypt
pyfiglet
instaloader
Pillow
